backend/requirements.txt
fastapi
pydantic
langchain-huggingface
langchain-qdrant
qdrant-client
langchain-groq
uvicorn
python-dotenv